+++ library/lib/src/image_size_getter.dart
@@ -1,145 +1,134 @@
-import 'package:collection/collection.dart';
+import 'dart:collection';
+
 import 'package:image_size_getter/image_size_getter.dart';
-import 'package:image_size_getter/src/decoder/sync/decoder.dart';
-import 'package:image_size_getter/src/decoder/async/decoder.dart' as ad;
 
 export 'core/input.dart';
 
-class ImageSizeGetter {
-  static bool isJpg(ImageInput input) {
-    if (!input.exists()) {
-      return false;
+/// {@template image_size_getter._DecoderContainer}
+///
+/// [_DecoderContainer] is a container for [BaseDecoder]s.
+///
+/// {@endtemplate}
+class _DecoderContainer extends IterableBase<BaseDecoder> {
+  /// {@macro image_size_getter._DecoderContainer}
+  _DecoderContainer(List<BaseDecoder> decoders) {
+    for (final decoder in decoders) {
+      _decoders[decoder.decoderName] = decoder;
     }
-
-    const start = [0xFF, 0xD8];
-    const end = [0xFF, 0xD9];
-
-    final length = input.length;
-    final startList = input.getRange(0, 2);
-    final endList = input.getRange(length - 2, length);
-
-    const eq = ListEquality();
-
-    return eq.equals(start, startList) && eq.equals(end, endList);
   }
 
-  static bool isPng(ImageInput input) {
-    final length = input.length;
+  /// The [BaseDecoder]s.
+  final Map<String, BaseDecoder> _decoders = {};
 
-    final start = input.getRange(0, 8);
-    final end = input.getRange(length - 12, length);
-    const eq = IterableEquality();
-    if (eq.equals(start, _PngHeaders.sig) && eq.equals(end, _PngHeaders.iend)) {
-      return true;
-    }
-
-    return false;
+  /// {@template image_size_getter._DecoderContainer.register}
+  ///
+  /// Registers a [BaseDecoder] to the container.
+  ///
+  /// If the [BaseDecoder] is already registered, it will be replaced.
+  ///
+  /// {@endtemplate}
+  void registerDecoder(BaseDecoder decoder) {
+    _decoders[decoder.decoderName] = decoder;
   }
 
-  static bool isWebp(ImageInput input) {
-    final sizeStart = input.getRange(0, 4);
-    final sizeEnd = input.getRange(8, 12);
-
-    const eq = ListEquality();
-
-    if (eq.equals(sizeStart, _WebpHeaders.fileSizeStart) &&
-        eq.equals(sizeEnd, _WebpHeaders.fileSizeEnd)) {
-      return true;
-    }
-    return false;
-  }
-
-  static bool isGif(ImageInput input) {
-    const eq = ListEquality();
-    final length = input.length;
-
-    final sizeStart = input.getRange(0, 6);
-    final sizeEnd = input.getRange(length - 1, length);
-
-    return eq.equals(sizeStart, _GifHeaders.start) &&
-        eq.equals(sizeEnd, _GifHeaders.end);
-  }
-
-  static Size getSize(ImageInput input) {
-    if (!input.exists()) {
-      throw Exception('The input is not exists.');
-    }
-
-    if (isJpg(input)) {
-      return JpegDecoder(input).size;
-    }
-    if (isPng(input)) {
-      return PngDecoder(input).size;
-    }
-    if (isWebp(input)) {
-      return WebpDecoder(input).size;
-    }
-    if (isGif(input)) {
-      return GifDecoder(input).size;
-    }
-
-    throw Exception('The input is not supported.');
-  }
+  @override
+  Iterator<BaseDecoder> get iterator => _decoders.values.iterator;
 }
 
-class AsyncImageSizeGetter {
-  static Future<bool> isJpg(AsyncImageInput input) async {
-    if (!(await input.exists())) {
-      return false;
+/// The instance of [_DecoderContainer].
+///
+/// This instance is used to register [BaseDecoder]s, it will be used by [ImageSizeGetter].
+final _decoders = _DecoderContainer([
+  const GifDecoder(),
+  const JpegDecoder(),
+  const WebpDecoder(),
+  const PngDecoder(),
+  const BmpDecoder(),
+]);
+
+/// {@template image_size_getter.ImageSizeGetter}
+///
+/// The main class of [ImageSizeGetter].
+///
+/// Simple example:
+///
+/// ```dart
+/// import 'dart:io';
+///
+/// import 'package:image_size_getter/image_size_getter.dart';
+/// import 'package:image_size_getter/file_input.dart'; // For compatibility with flutter web.
+///
+/// void main(List<String> arguments) async {
+///   final file = File('asset/IMG_20180908_080245.jpg');
+///   final size = ImageSizeGetter.getSize(FileInput(file));
+///   print('jpg size: $size');
+/// }
+/// ```
+///
+/// {endtemplate}
+class ImageSizeGetter {
+  /// {@macro image_size_getter._DecoderContainer.register}
+  static void registerDecoder(BaseDecoder decoder) {
+    _decoders.registerDecoder(decoder);
+  }
+
+  /// Returns the [input] is png format or not.
+  ///
+  /// See also: [PngDecoder.isValid] or [PngDecoder.isValidAsync].
+  static bool isPng(ImageInput input) {
+    return PngDecoder().isValid(input);
+  }
+
+  /// Returns the [input] is webp format or not.
+  ///
+  /// See also: [WebpDecoder.isValid] or [WebpDecoder.isValidAsync].
+  static bool isWebp(ImageInput input) {
+    return WebpDecoder().isValid(input);
+  }
+
+  /// Returns the [input] is gif format or not.
+  ///
+  /// See also: [GifDecoder.isValid] or [GifDecoder.isValidAsync].
+  static bool isGif(ImageInput input) {
+    return GifDecoder().isValid(input);
+  }
+
+  /// Returns the [input] is jpeg format or not.
+  ///
+  /// See also: [JpegDecoder.isValid] or [JpegDecoder.isValidAsync].
+  static bool isJpg(ImageInput input) {
+    return JpegDecoder().isValid(input);
+  }
+
+  /// {@template image_size_getter.getSize}
+  ///
+  /// Get the size of the [input].
+  ///
+  /// If the [input] not exists, it will throw [StateError].
+  ///
+  /// If the [input] is not a valid image format, it will throw [UnsupportedError].
+  ///
+  /// {@endtemplate}
+  static Size getSize(ImageInput input) {
+    if (!input.exists()) {
+      throw StateError('The input is not exists.');
     }
 
-    const start = [0xFF, 0xD8];
-    const end = [0xFF, 0xD9];
-
-    final length = await input.length;
-    final startList = await input.getRange(0, 2);
-    final endList = await input.getRange(length - 2, length);
-
-    const eq = ListEquality();
-
-    return eq.equals(start, startList) && eq.equals(end, endList);
-  }
-
-  static Future<bool> isPng(AsyncImageInput input) async {
-    final length = await input.length;
-
-    final start = await input.getRange(0, 8);
-    final end = await input.getRange(length - 12, length);
-    const eq = IterableEquality();
-    if (eq.equals(start, _PngHeaders.sig) && eq.equals(end, _PngHeaders.iend)) {
-      return true;
+    for (var value in _decoders) {
+      if (value.isValid(input)) {
+        return value.getSize(input);
+      }
     }
 
-    return false;
+    throw UnsupportedError('The input is not supported.');
   }
 
-  static Future<bool> isWebp(AsyncImageInput input) async {
-    final sizeStart = await input.getRange(0, 4);
-    final sizeEnd = await input.getRange(8, 12);
-
-    const eq = ListEquality();
-
-    if (eq.equals(sizeStart, _WebpHeaders.fileSizeStart) &&
-        eq.equals(sizeEnd, _WebpHeaders.fileSizeEnd)) {
-      return true;
-    }
-    return false;
-  }
-
-  static Future<bool> isGif(AsyncImageInput input) async {
-    const eq = ListEquality();
-    final length = await input.length;
-
-    final sizeStart = await input.getRange(0, 6);
-    final sizeEnd = await input.getRange(length - 1, length);
-
-    return eq.equals(sizeStart, _GifHeaders.start) &&
-        eq.equals(sizeEnd, _GifHeaders.end);
-  }
-
-  static Future<Size> getSize(AsyncImageInput input) async {
+  /// {@macro image_size_getter.getSize}
+  ///
+  /// The method is async.
+  static Future<Size> getSizeAsync(AsyncImageInput input) async {
     if (!await input.exists()) {
-      throw Exception('The input is not exists.');
+      throw StateError('The input is not exists.');
     }
 
     if (!(await input.supportRangeLoad())) {
@@ -151,76 +140,12 @@ class AsyncImageSizeGetter {
       }
     }
 
-    if (await isJpg(input)) {
-      return ad.JpegDecoder(input).size;
-    }
-    if (await isPng(input)) {
-      return ad.PngDecoder(input).size;
-    }
-    if (await isWebp(input)) {
-      return ad.WebpDecoder(input).size;
-    }
-    if (await isGif(input)) {
-      return ad.GifDecoder(input).size;
+    for (var value in _decoders) {
+      if (await value.isValidAsync(input)) {
+        return value.getSizeAsync(input);
+      }
     }
 
-    throw Exception('The input is not supported.');
+    throw UnsupportedError('The input is not supported.');
   }
-}
-
-class _PngHeaders {
-  static const sig = [
-    0x89,
-    0x50,
-    0x4E,
-    0x47,
-    0x0D,
-    0x0A,
-    0x1A,
-    0x0A,
-  ];
-
-  static const iend = [
-    0x00,
-    0x00,
-    0x00,
-    0x00,
-    0x49,
-    0x45,
-    0x4E,
-    0x44,
-    0xAE,
-    0x42,
-    0x60,
-    0x82
-  ];
-}
-
-class _WebpHeaders {
-  static const fileSizeStart = [
-    0x52,
-    0x49,
-    0x46,
-    0x46,
-  ];
-
-  static const fileSizeEnd = [
-    0x57,
-    0x45,
-    0x42,
-    0x50,
-  ];
-}
-
-class _GifHeaders {
-  static const start = [
-    0x47,
-    0x49,
-    0x46,
-    0x38,
-    0x39,
-    0x61,
-  ];
-
-  static const end = [0x3B];
 }